clean up docs and artifacts
