+++ CLAUDE.md
@@ -44,7 +44,7 @@ Extends `ss.Disease` from Starsim.
 
 ### Multi-strain / drug-resistance extension
 
-[tbsim/resistance/](tbsim/resistance/) adds a multi-strain, drug-resistance overlay. `TBResistant` (in [tb_resistant.py](tbsim/resistance/tb_resistant.py)) is a drop-in `TB` subclass that encodes `2ⁿ` strains over `n` named drugs as a per-agent bitmask (`strain_mask`) plus a per-strain multiplicity counter, with per-strain fitness costs, superinfection, and de-novo resistance. It ships strain-aware product/delivery interventions — `TxR`/`TxDeliveryR` (treatment), `DST`/`DSTDelivery` (diagnostics), `TPTRx` (preventive therapy) — the `ResistanceStats`/`StrainResults` analyzers, and a two-strain reference ODE (`tbsim.compartmental.TwoStrainODE`) for ABM↔ODE validation. All public entry points are re-exported on `tbsim`. CI coverage lives in [tests/test_resistance.py](tests/test_resistance.py); the broader dev suite is `tbsim/resistance/devtests/` (run manually, not collected by CI). Docs (user manual, updates guide, tech spec) are under `tbsim/resistance/docs/`.
+[tbsim/resistance/](tbsim/resistance/) adds a multi-strain, drug-resistance overlay. `TBResistant` (in [tb_resistant.py](tbsim/resistance/tb_resistant.py)) is a drop-in `TB` subclass that encodes `2ⁿ` strains over `n` named drugs as a per-agent bitmask (`strain_mask`) plus a per-strain multiplicity counter, with per-strain fitness costs, superinfection, and de-novo resistance. It ships strain-aware product/delivery interventions — `TxR`/`TxDeliveryR` (treatment), `DST`/`DSTDelivery` (diagnostics), `TPTRx` (preventive therapy) — the `ResistanceStats`/`StrainResults` analyzers, and a two-strain reference ODE (`tbsim.compartmental.TwoStrainODE`) for ABM↔ODE validation. All public entry points are re-exported on `tbsim`. CI coverage lives in [tests/test_resistance.py](tests/test_resistance.py) and the ABM↔ODE validation in [tests/test_compartmental.py](tests/test_compartmental.py). The implementation reference is [tbsim/resistance/README.md](tbsim/resistance/README.md); the usage tutorial is `docs/tutorials/resistance_tutorial.qmd`.
 
 ### Sim Wrapper
