+++ docs/tutorials/resistance_tutorial.qmd
@@ -9,7 +9,7 @@ strains, multi-strain transmission and superinfection, competition and reinfecti
 and treatment-acquired resistance, drug-susceptibility testing (DST) with regimen routing,
 treatment monitoring, and strain-aware preventive therapy (TPT).
 
-The companion reference — [tbsim-resistance-implementation.md](https://github.com/starsimhub/tbsim/blob/main/tbsim/resistance/docs/tbsim-resistance-implementation.md) — explains *how* each piece is implemented; this tutorial focuses on *using* it.
+The companion reference — the [`tbsim.resistance` README](https://github.com/starsimhub/tbsim/blob/main/tbsim/resistance/README.md) — explains *how* each piece is implemented; this tutorial focuses on *using* it.
 
 ## Setup
 
@@ -219,6 +219,28 @@ print(f"first-line courses:      {int(np.sum(sim.results['first'].n_treated))}")
 print(f"second-line (RIF-R):     {int(np.sum(sim.results['second'].n_treated))}")
 ```
 
+### Retreatment vs. new cases
+
+Every `TxDeliveryR` stamps a durable, cross-regimen `tb.ti_last_treatment` when it starts a course, so a
+delivery can distinguish agents returning after a previous course (retreatment) from genuinely new cases.
+`TxDeliveryR.failure_case_eligibility(within=<duration>)` returns a `sim → uids` classifier selecting
+agents whose last treatment was within `within` (pass `new_case=True` for the complement) — the natural
+way to route retreatment cases to a stronger second-line regimen.
+
+```{python}
+tb = tbsim.TBResistant(drugs=['RIF'], rel_fitness={'RIF': 0.9},
+                       beta=ss.permonth(0.35), init_prev=ss.bernoulli(0.12),
+                       init_strains=[0.8, 0.2], rr_reinfection_inf=1.0, rr_reinfection_non=1.0)
+retreat = tbsim.TxDeliveryR.failure_case_eligibility(within=ss.years(2))  # returned to care within 2 years
+first = tbsim.TxDeliveryR(name='first', rate_sym=ss.peryear(1.0),
+                          product=tbsim.TxR(strains=tb.strains, base_efficacy=0.5, resist_penalty={'RIF': 0.2}))
+second = tbsim.TxDeliveryR(name='second', eligibility=retreat, supersedes=['first'], rate_sym=ss.peryear(1.0),
+                           product=tbsim.TxR(strains=tb.strains, base_efficacy=0.85, regimen_drugs=['RIF']))
+sim = build_sim(tb, interventions=[first, second], stop='2035-12-31'); sim.run()
+print(f"first-line courses:  {int(np.sum(sim.results['first'].n_treated))}")
+print(f"retreatment courses: {int(np.sum(sim.results['second'].n_treated))}")
+```
+
 ## 7. Treatment monitoring and regimen switching
 
 `treatment_monitoring_eligibility(tx_name, after_steps)` selects agents who have been on a course for
@@ -315,5 +337,78 @@ You have now used every major feature of the resistance extension:
 - **`TPTRx`** — strain-aware preventive therapy and the resistance-unmasking dynamic.
 - **`ResistanceStats` / `StrainResults`** — resistance-origin decomposition and per-strain tracking.
 
-For the mechanics behind each of these, see the implementation reference,
-`tbsim/resistance/docs/tbsim-resistance-implementation.md`.
+For the mechanics behind each of these, see the [`tbsim.resistance` README](https://github.com/starsimhub/tbsim/blob/main/tbsim/resistance/README.md).
+
+## Reference
+
+### Results recorded by `TBResistant`
+
+Alongside the standard TB outputs, `TBResistant` records:
+
+| Result | Meaning |
+|--------|---------|
+| `frac_resist` | Fraction of active TB with any resistance |
+| `frac_resist_<drug>` | Fraction resistant to that drug (aggregate phenotype) |
+| `frac_super` | Fraction of active TB that is superinfected |
+| `new_denovo_resistance` | De-novo acquisition events this step |
+| `new_transmitted_resistance` | New acquisitions of a resistant strain via transmission |
+| `new_identical_superinf` | Identical-strain re-exposures this step (each increments the agent's per-strain count) |
+
+### Parameter cheat sheet
+
+**`TBResistant` / `Strains`**
+
+| Parameter | Type | Default | Notes |
+|-----------|------|---------|-------|
+| `drugs` | list[str] | `['TX']` | Ordered drug names; `m = 2**len(drugs)` |
+| `rel_fitness` | dict | `{}` | Per-drug `r_i ∈ [0,1]`; missing → 1.0 |
+| `rr_reinfection_inf` | float | → `rr_reinfection_rec` | σ for INFECTION |
+| `rr_reinfection_non` | float | → `rr_reinfection_inf` | σ for NON_INFECTIOUS |
+| `rr_reinfection_asy` / `_sym` | float | `0` | Superinfection in active disease |
+| `p_multi` | float | `1` | Prob. keep all strains at →ASY |
+| `prog_select` | str | `'random'` | Or `'fitness'` under bottleneck |
+| `rr_prog_super` / `rr_clear_super` | float | `1` | Optional multi-strain rate multipliers |
+| `p_rand` | dict | off | De-novo `{drug: p}` |
+| `prog_resist_mode` | str | `'mixed'` | Or `'replacement'` |
+| `init_strains` | array | pan only | Seed mix over strain ids |
+
+**`TxR` / `TxDeliveryR`**
+
+| Parameter | Notes |
+|-----------|--------|
+| `base_efficacy` | Cure prob. for a fully susceptible strain |
+| `resist_penalty` | Per-drug multiplier on efficacy for resistance to **regimen** drugs |
+| `efficacy_by_strain` | Explicit per-strain cure-prob vector (length `m`); overrides `base_efficacy`/`resist_penalty` |
+| `adherence` | Per-course completion prob. correlating an agent's strains; **float** (shared) or **callable** `uids → prob` |
+| `q_acq` | Per-drug acquisition-on-failure (replacement) |
+| `acq_state_rr` | Scale `q_acq` by TB state at failure |
+| `regimen_drugs` | Which drugs the regimen contains |
+| `rate_asym` / `rate_sym` | Initiation rates |
+| `dur_treatment` | Course length before the outcome resolves (default `ss.months(6)`) |
+| `eligibility` | Optional `sim → uids` override |
+| `supersedes` | Names of deliveries to interrupt before starting |
+| `retreat_after` | Refractory `ss.dur` after a course before the same agent is re-treated by this delivery |
+| `treat_latent` | `False` (default) = certain strain-aware sterilization of a selected latent agent (not counted in `n_treated`); `True` = full failable course |
+| `failure_case_eligibility(within, base=, new_case=)` | Classifier `sim → uids` for retreatment vs new case (reads durable `tb.ti_last_treatment`) |
+
+**`DST` / `TPTRx`**
+
+| Parameter | Notes |
+|-----------|--------|
+| `sens` / `spec` (DST) | Scalar or per-drug dict |
+| `p_strain_obs` (DST) | `None` → use fitness; or scalar/dict |
+| `result_validity` (DSTDelivery) | Stored results older than this are wiped, forcing a re-test |
+| `DSTDelivery.matches(**drugs, max_age=)` | Eligibility from the observed profile (optionally within `max_age`) |
+| `regimen_drugs` (TPT) | Drugs that must all be susceptible for sterilization |
+| `p_tpt_acq` (TPT) | Acquisition among ineffective TPT outcomes |
+| `acq_state_rr` (TPT) | Scale `p_tpt_acq` by TB state |
+| `p_sterilize` (TPT) | Must be > 0 to exercise strain-aware clearance |
+
+### Known limitations
+
+| Topic | Current behavior |
+|-------|------------------|
+| DST indeterminate | Binary observed profile only (no explicit indeterminate outcome) |
+| TPT partial efficacy | Sterilization is all-or-nothing per strain; no per-drug TPT `resist_penalty` |
+| Time-varying progression hazard | Optional exponential decline via `k_asy`/`k_non` (defaults 0 = constant hazard) |
+| LTFU outcome | Not modeled as a separate treatment outcome |
