+++ docs/user_guide.md
@@ -95,6 +95,18 @@ TBsim includes sophisticated connectors to link TB with other diseases:
 - **Nutrition-TB Connector**: Nutritional status impacts TB progression and treatment
 - **Supplementation Effects**: Nutritional intervention modeling
 
+## Drug Resistance and Multi-Strain TB
+
+TBsim includes an optional multi-strain, drug-resistance overlay (`tbsim.resistance`) built on the same agent-level natural history. `TBResistant` is a drop-in replacement for `TB` that encodes `2ⁿ` circulating strains over `n` named drugs as a per-agent bitmask, with:
+
+- **Per-drug fitness costs** — each drug carries a transmission fitness penalty, so resistant strains can be out-competed unless selected for.
+- **Superinfection and competition** — agents can carry multiple strains; a transmission/progression bottleneck governs which strain moves.
+- **De-novo and treatment-acquired resistance** — resistance can emerge by mutation at progression or on treatment failure, per drug.
+- **Strain-aware interventions** — treatment (`TxR` / `TxDeliveryR`), drug-susceptibility testing with regimen routing (`DST` / `DSTDelivery`), and preventive therapy (`TPTRx`) all resolve at the strain level.
+- **Origin decomposition** — the `ResistanceStats` analyzer attributes new resistant infections to de-novo, treatment-acquired, transmitted, and TPT-acquired channels.
+
+The agent-based model is validated against a deterministic two-strain reference ODE (`tbsim.compartmental.TwoStrainODE`). For a hands-on, feature-by-feature walkthrough, see the [drug resistance tutorial](tutorials/resistance_tutorial.qmd).
+
 ## Networks and Social Structure
 
 **Network Types**
