--- tbsim/resistance/docs/update-v2-notes.md
@@ -1,13 +1,0 @@
-# Update v2 notes
-
-## Reinfection
-
-I've started going through the time-varying progression stuff/PRs 441 and 442. The only issue I've noticed is that it looks like INFECTED agents are not currently "eligible" for reinfection in base TB (although this was added for TBresistant). This means that we don't allow INFECTED agents to get their clocks reset (implying higher progression) when they are re-exposed/re-infected. What we would like is for this behavior to mirror TBresistant (we allow for rr_reinfection_inf which can equal rr_reinfection_rec as a default, and INFECTED agents who are reinfected get ti_infected reset to the current timestep and are thus at higher risk of progression - just no need to track any strains/strain counts).
-
-## TR Notes July 15
-
-1. Treatment-acquired resistance: Right now, only 1 strain can acquire resistance to each drug per treatment round. This is no longer always the "first"/"most susceptible" strain (good!) but we want to allow for the possibility of multiple strains acquiring resistance (similar to how spontaneous resistance is handled). 
-2. Strain counts + resistance acquisition: If a strain with count 2 acquires resistance (whether spontaneously, through treatment, or through TPT), the model currently removes that strain from an agent’s strain profile (correct) but the updated now-resistant strain gets its count reset at 1. We would like to maintain the count (i.e., count stays at 2 in this example)
-3. Strain counts + resistance acquisition: As an extension of #2, if this agent already had this resistant strain in their strain profile (whether with count 1 or > 1), that count stays the same in the current model - it does not get added to. This is equivalent behavior to clearing the newly resistant strain. We would instead like the resistant strain’s count to be increased (for example, if the count had been 2, and 2 susceptible strains acquired new resistance, the count should now be 4). This applies to all 3 mechanisms of resistance acquisition (de novo/spontaneous, tx-acquired, TPT-acquired)
-4. DST: The DST product draws a single random number per strain and reuses it across all drugs, so within one strain that is resistant to some modeled drugs and susceptible to others, the test's errors are correlated across drugs rather than independent — a genuinely RIF-resistant/BDQ-susceptible case can be read correctly or "swapped," but essentially never over-called as resistant to both. This doesn't affect any single drug's sensitivity/specificity (those marginals are exact) and is inert unless treatment is routed on the joint multi-drug DST profile under imperfect testing. It seems that this can be addressed for now by modeling separate DST products for each drug (so we can use this as a workaround), but would be good to fix as some DST products genuinely do test for multiple drugs. 
-5. Treatment of latent/INFECTED agents: INFECTED agents treated for active disease with a treatment product are either all sterilized/cleared (treat_latent=FALSE) or have similar behavior (including regimen efficacy and resistance acquisition) as agents with active disease (treat_latent=TRUE). What we probably want here is for all INFECTED agents to be cleared unless they have prior resistance, with no additional resistance acquisition. As a workaround for now, it seems that this can currently be accomplished by creating separate treatment products with eligibility rules such that active agents only get 1 product and INFECTED agents only get the other.
